+++ admin/src/PreferenceMaster/LeaguesPreference.js
@@ -15,7 +15,7 @@ import 'react-toastify/dist/ReactToastify.css';
 import { useState, useEffect } from "react";
 import useForm from "../useForm";
 import IconButton from '@mui/material/IconButton';
-
+import Swal from 'sweetalert2'
 
 export default function LeaguesPreference() {
 
@@ -69,29 +69,46 @@ export default function LeaguesPreference() {
             })
     }
 
-    /////////////////delete complaint /////////////////
-    const deleteCategory = (_id) => {
 
-        axios.delete(`/web_api/leagues/${_id}`)
+
+  /////////////////delete api call /////////////////
+   const deleteCategory = (_id) => {
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {         
+            axios.delete(`/web_api/leagues/${_id}`)
             .then(res => {
                 if (res.status) {
                     let data = res.data;
-
-                    if (data.status) {
-                        toast.success(data.msg);
-                        return PreferenceList();
+                    if (data.status) { 
+                        Swal.fire(
+                            'Deleted!',
+                             data.msg,
+                            'success'
+                          )
+                         return PreferenceList();
                     } else {
-                        toast.error('something went wrong please try again');
+                        toast.error(data.msg);
                     }
                 }
                 else {
-                    toast.error('something went wrong please try again..');
+                    toast.error(data.msg);
                 }
             })
-            .catch(error => {
-                console.log(this.state);
-            })
-    }
+        }
+      })
+}
+
+
+
+
 
     /////////////////complaint list/////////////////
     const onCloseModal = () => setOpen(false);
@@ -163,7 +180,7 @@ export default function LeaguesPreference() {
                         let data = response.data;
                         console.log(data.msg)
                         if (data.status) {
-                            toast.success(data.msg);
+                            toast.success("League Added Successfully");
                             e.target.reset();
                             return PreferenceList();
                         } else {
